+++ CHANGELOG.md
@@ -8,6 +8,7 @@ This file documents all changes to Argus-frontend. This file is primarily meant 
 - Fix a bug in Timeslot's time pickers where changing start-/end time via icon, leads to wrong values being registered, misleading error text being displayed, and "Save"-button being disabled 
 - Fix a bug in Timeslot's time pickers where typing invalid values for both start- and end time (both input fields display error text), disrupts further changes (incl. corrections) of those time values
 - Fix a bug in Timeslot's time pickers where error text is not always removed and "Save"-button is disabled, even after user has provided valid and non-conflicting values (time range where start time is before end time) 
+- Fix bug where incidents count was messed up due to user switching between table pages too rapidly
 
 - Labeling of "Unacked" chips and buttons is now consistent.
 - Incidents page does not blank out anymore when selecting an old-style filter (filter created before 2023).